Move job logs from health to jobs router
